Added everything up to and including sending configuration to the database

+++ src/main/java/org/erau/eas/serverweb/Repository/FlightsRepository.java
@@ -3,5 +3,5 @@ package org.erau.eas.serverweb.Repository;
 import org.erau.eas.serverweb.db.Flights;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FlightsRepository extends CrudRepository<Flights, Long>{
+public interface FlightsRepository extends CrudRepository<Flights, Integer>{
 }
